+++ Sources/Main/Util/Private/AutomaticFieldsBuilder.swift
@@ -139,11 +139,15 @@ final class AutomaticFieldsBuilder: AutomaticFieldsBuildable {
         // MARK: mnetwd
         payload[PayloadParameterKeys.Telephony.mnetwd] = telephonyHandler.mnetwd ?? ""
 
-        // MARK: netop
-        payload[PayloadParameterKeys.Telephony.netop] = telephonyHandler.netop
+        // MARK: simop
+        if !telephonyHandler.simop.isEmpty {
+            payload[PayloadParameterKeys.Telephony.simop] = telephonyHandler.simop
+        }
 
-        // MARK: netopn
-        payload[PayloadParameterKeys.Telephony.netopn] = telephonyHandler.netopn
+        // MARK: simopn
+        if !telephonyHandler.simopn.isEmpty {
+            payload[PayloadParameterKeys.Telephony.simopn] = telephonyHandler.simopn
+        }
 
         // MARK: mori
         payload[PayloadParameterKeys.Orientation.mori] = NSNumber(value: statusBarOrientationHandler.mori.rawValue)