+++ Tests/GeoSpec/GeoTrackerSpec.swift
@@ -273,9 +273,9 @@ final class GeoTrackerSpec: QuickSpec {
 
                         expect(json[PayloadParameterKeys.Telephony.mnetwd] as? Int).to(equal(1))
 
-                        expect(json[PayloadParameterKeys.Telephony.netopn] as? String).to(beEmpty())
+                        expect(json[PayloadParameterKeys.Telephony.simopn] as? String).to(beNil())
 
-                        expect(json[PayloadParameterKeys.Telephony.netop] as? String).to(beEmpty())
+                        expect(json[PayloadParameterKeys.Telephony.simop] as? String).to(beNil())
 
                         expect(json[PayloadParameterKeys.Device.model] as? String).to(equal(expectedModel))