+++ CHANGELOG.md
@@ -41,6 +41,7 @@
 * [CONRAT-28925](https://jira.rakuten-it.com/jira/browse/CONRAT-28925): Fix missing action parameters on request location.
 * [CONRAT-28848](https://jira.rakuten-it.com/jira/browse/CONRAT-28848): Handle background timer calculation for GeoTracker.
 * [CONRAT-29004](https://jira.rakuten-it.com/jira/browse/CONRAT-29004): Fix configured time interval functionality for start/stop location collection.
+* [CONRAT-290331](https://jira.rakuten-it.com/jira/browse/CONRAT-29331): Do not send empty simopn and simop.
 
 ### Changes
 * [CONRAT-27817](https://jira.rakuten-it.com/jira/browse/CONRAT-27817): Remove dependency on RSDKUtils.